[rust] Allow developers to develop Rust functions on Windows
