+++ crates/vercel_runtime/Cargo.toml
@@ -18,12 +18,14 @@ tokio = { version = "1", features = ["full"] }
 http-body-util = "0.1"
 tokio-stream = "0.1"
 lazy_static = "1.4"
-libc = "0.2"
 http-body = "1.0"
 tower = { version = "0.5", features = ["util"] }
 
 axum = { version = "0.8", optional = true }
 axum-streams = { version = "0.24", features = ["json", "text"], optional = true }
+
+[target.'cfg(unix)'.dependencies]
+libc = "0.2"
 
 [features]
 default = []
